refactor(selector-v2): flatten Selectors dir to repo convention
Move business selectors from src/renderer/src/components/selectors/v2/
to src/renderer/src/components/Selectors/. PascalCase top-level matches
Popups/Buttons/Tags; the v2/ subdirectory is unnecessary nesting.
Filenames keep the V2 suffix as a migration marker alongside v1
ModelSelector.
