+++ src/renderer/src/components/Selectors/AssistantSelectorV2.tsx
