+++ src/renderer/src/components/Selectors/index.ts
